Add DICharts handler & extract DICharts class to own file
