+++ di_website/visualisation/static/visualisation/js/dicharts.js
@@ -1,1 +1,1 @@
-var DICharts=function(e){var t={};function r(n){if(t[n])return t[n].exports;var o=t[n]={i:n,l:!1,exports:{}};return e[n].call(o.exports,o,o.exports,r),o.l=!0,o.exports}return r.m=e,r.c=t,r.d=function(e,t,n){r.o(e,t)||Object.defineProperty(e,t,{enumerable:!0,get:n})},r.r=function(e){"undefined"!=typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(e,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(e,"__esModule",{value:!0})},r.t=function(e,t){if(1&t&&(e=r(e)),8&t)return e;if(4&t&&"object"==typeof e&&e&&e.__esModule)return e;var n=Object.create(null);if(r.r(n),Object.defineProperty(n,"default",{enumerable:!0,value:e}),2&t&&"string"!=typeof e)for(var o in e)r.d(n,o,function(t){return e[t]}.bind(null,o));return n},r.n=function(e){var t=e&&e.__esModule?function(){return e.default}:function(){return e};return r.d(t,"a",t),t},r.o=function(e,t){return Object.prototype.hasOwnProperty.call(e,t)},r.p="",r(r.s=1)}([function(e,t,r){"use strict";var n=function(e){return function(e){return!!e&&"object"==typeof e}(e)&&!function(e){var t=Object.prototype.toString.call(e);return"[object RegExp]"===t||"[object Date]"===t||function(e){return e.$$typeof===o}(e)}(e)};var o="function"==typeof Symbol&&Symbol.for?Symbol.for("react.element"):60103;function a(e,t){return!1!==t.clone&&t.isMergeableObject(e)?f((r=e,Array.isArray(r)?[]:{}),e,t):e;var r}function i(e,t,r){return e.concat(t).map((function(e){return a(e,r)}))}function c(e){return Object.keys(e).concat(function(e){return Object.getOwnPropertySymbols?Object.getOwnPropertySymbols(e).filter((function(t){return e.propertyIsEnumerable(t)})):[]}(e))}function u(e,t){try{return t in e}catch(e){return!1}}function l(e,t,r){var n={};return r.isMergeableObject(e)&&c(e).forEach((function(t){n[t]=a(e[t],r)})),c(t).forEach((function(o){(function(e,t){return u(e,t)&&!(Object.hasOwnProperty.call(e,t)&&Object.propertyIsEnumerable.call(e,t))})(e,o)||(u(e,o)&&r.isMergeableObject(t[o])?n[o]=function(e,t){if(!t.customMerge)return f;var r=t.customMerge(e);return"function"==typeof r?r:f}(o,r)(e[o],t[o],r):n[o]=a(t[o],r))})),n}function f(e,t,r){(r=r||{}).arrayMerge=r.arrayMerge||i,r.isMergeableObject=r.isMergeableObject||n,r.cloneUnlessOtherwiseSpecified=a;var o=Array.isArray(t);return o===Array.isArray(e)?o?r.arrayMerge(e,t,r):l(e,t,r):a(t,r)}f.all=function(e,t){if(!Array.isArray(e))throw new Error("first argument should be an array");return e.reduce((function(e,r){return f(e,r,t)}),{})};var s=f;e.exports=s},function(e,t,r){"use strict";r.r(t),r.d(t,"Chart",(function(){return f}));var n=r(0),o=r.n(n);function a(e,t,r){return t in e?Object.defineProperty(e,t,{value:r,enumerable:!0,configurable:!0,writable:!0}):e[t]=r,e}var i={displayModeBar:!0,responsive:!0,showLink:!1,plotlyServerURL:"https://chart-studio.plotly.com",toImageButtonOptions:{width:1200,height:657},modeBarButtonsToRemove:["pan2d","select2d","lasso2d","hoverClosestCartesian","hoverCompareCartesian","toggleSpikelines"]},c=function e(t,r){var n=this;!function(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}(this,e),a(this,"plotly",void 0),a(this,"chartElement",void 0),a(this,"getPlotly",(function(){return n.plotly})),a(this,"getChartElement",(function(){return n.chartElement})),a(this,"addChartElement",(function(e){return n.chartElement=e,n})),a(this,"getConfig",(function(){var e=arguments.length>0&&void 0!==arguments[0]?arguments[0]:{};return o()(i,e)})),this.plotly=t,this.chartElement=r};function u(e,t){for(var r=0;r<t.length;r++){var n=t[r];n.enumerable=n.enumerable||!1,n.configurable=!0,"value"in n&&(n.writable=!0),Object.defineProperty(e,n.key,n)}}function l(e,t,r){return t in e?Object.defineProperty(e,t,{value:r,enumerable:!0,configurable:!0,writable:!0}):e[t]=r,e}var f=function(){function e(t){var r=this;!function(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}(this,e),l(this,"chartElement",void 0),l(this,"showLoading",(function(){var e=r.getParentElement();e&&e.classList.add("chart-container--loading")})),l(this,"hideLoading",(function(){var e=r.getParentElement();e&&function(e){e.classList.remove("chart-container--loading");var t=e.querySelector(".chart-loading");t&&t.classList.add("hide")}(e)})),l(this,"addFilter",(function(e,t,n){var o=r.getElement(e);return t.forEach((function(e){var t=document.createElement("option");t.value=e,t.text=r.getFilterText(e,n),o.appendChild(t)})),o.classList.add("data-selector--active"),o})),l(this,"addPlotly",(function(e){return new c(e,r.chartElement)})),this.chartElement=this.getElement(t)}var t,r,n;return t=e,(r=[{key:"getElement",value:function(e){if(!e)throw new Error("An id to the element or an actual element is required");if("string"==typeof e){var t=document.getElementById(e);if(!t)throw new Error("No element with id ".concat(e));return t}return e}},{key:"getParentElement",value:function(){return this.chartElement.parentElement}},{key:"getFilterText",value:function(e){var t=arguments.length>1&&void 0!==arguments[1]?arguments[1]:{},r=t.labelPrefix,n=t.labelSuffix;return"".concat(r||"").concat(e).concat(n||"")}}])&&u(t.prototype,r),n&&u(t,n),e}()}]);
+var DICharts=function(e){var t={};function r(n){if(t[n])return t[n].exports;var o=t[n]={i:n,l:!1,exports:{}};return e[n].call(o.exports,o,o.exports,r),o.l=!0,o.exports}return r.m=e,r.c=t,r.d=function(e,t,n){r.o(e,t)||Object.defineProperty(e,t,{enumerable:!0,get:n})},r.r=function(e){"undefined"!=typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(e,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(e,"__esModule",{value:!0})},r.t=function(e,t){if(1&t&&(e=r(e)),8&t)return e;if(4&t&&"object"==typeof e&&e&&e.__esModule)return e;var n=Object.create(null);if(r.r(n),Object.defineProperty(n,"default",{enumerable:!0,value:e}),2&t&&"string"!=typeof e)for(var o in e)r.d(n,o,function(t){return e[t]}.bind(null,o));return n},r.n=function(e){var t=e&&e.__esModule?function(){return e.default}:function(){return e};return r.d(t,"a",t),t},r.o=function(e,t){return Object.prototype.hasOwnProperty.call(e,t)},r.p="",r(r.s=1)}([function(e,t,r){"use strict";var n=function(e){return function(e){return!!e&&"object"==typeof e}(e)&&!function(e){var t=Object.prototype.toString.call(e);return"[object RegExp]"===t||"[object Date]"===t||function(e){return e.$$typeof===o}(e)}(e)};var o="function"==typeof Symbol&&Symbol.for?Symbol.for("react.element"):60103;function a(e,t){return!1!==t.clone&&t.isMergeableObject(e)?f((r=e,Array.isArray(r)?[]:{}),e,t):e;var r}function i(e,t,r){return e.concat(t).map((function(e){return a(e,r)}))}function c(e){return Object.keys(e).concat(function(e){return Object.getOwnPropertySymbols?Object.getOwnPropertySymbols(e).filter((function(t){return e.propertyIsEnumerable(t)})):[]}(e))}function u(e,t){try{return t in e}catch(e){return!1}}function l(e,t,r){var n={};return r.isMergeableObject(e)&&c(e).forEach((function(t){n[t]=a(e[t],r)})),c(t).forEach((function(o){(function(e,t){return u(e,t)&&!(Object.hasOwnProperty.call(e,t)&&Object.propertyIsEnumerable.call(e,t))})(e,o)||(u(e,o)&&r.isMergeableObject(t[o])?n[o]=function(e,t){if(!t.customMerge)return f;var r=t.customMerge(e);return"function"==typeof r?r:f}(o,r)(e[o],t[o],r):n[o]=a(t[o],r))})),n}function f(e,t,r){(r=r||{}).arrayMerge=r.arrayMerge||i,r.isMergeableObject=r.isMergeableObject||n,r.cloneUnlessOtherwiseSpecified=a;var o=Array.isArray(t);return o===Array.isArray(e)?o?r.arrayMerge(e,t,r):l(e,t,r):a(t,r)}f.all=function(e,t){if(!Array.isArray(e))throw new Error("first argument should be an array");return e.reduce((function(e,r){return f(e,r,t)}),{})};var s=f;e.exports=s},function(e,t,r){"use strict";r.r(t),r.d(t,"handler",(function(){return h})),r.d(t,"Chart",(function(){return d}));var n=r(0),o=r.n(n);function a(e,t,r){return t in e?Object.defineProperty(e,t,{value:r,enumerable:!0,configurable:!0,writable:!0}):e[t]=r,e}var i={displayModeBar:!0,responsive:!0,showLink:!1,plotlyServerURL:"https://chart-studio.plotly.com",toImageButtonOptions:{width:1200,height:657},modeBarButtonsToRemove:["pan2d","select2d","lasso2d","hoverClosestCartesian","hoverCompareCartesian","toggleSpikelines"]},c=function e(t,r){var n=this;!function(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}(this,e),a(this,"plotly",void 0),a(this,"chartElement",void 0),a(this,"getPlotly",(function(){return n.plotly})),a(this,"getChartElement",(function(){return n.chartElement})),a(this,"addChartElement",(function(e){return n.chartElement=e,n})),a(this,"getConfig",(function(){var e=arguments.length>0&&void 0!==arguments[0]?arguments[0]:{};return o()(i,e)})),this.plotly=t,this.chartElement=r};function u(e,t){for(var r=0;r<t.length;r++){var n=t[r];n.enumerable=n.enumerable||!1,n.configurable=!0,"value"in n&&(n.writable=!0),Object.defineProperty(e,n.key,n)}}function l(e,t,r){return t in e?Object.defineProperty(e,t,{value:r,enumerable:!0,configurable:!0,writable:!0}):e[t]=r,e}var f,s,d=function(){function e(t){var r=this;!function(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}(this,e),l(this,"chartElement",void 0),l(this,"showLoading",(function(){var e=r.getParentElement();e&&e.classList.add("chart-container--loading")})),l(this,"hideLoading",(function(){var e=r.getParentElement();e&&function(e){e.classList.remove("chart-container--loading");var t=e.querySelector(".chart-loading");t&&t.classList.add("hide")}(e)})),l(this,"addFilter",(function(e,t,n){var o=r.getElement(e);return t.forEach((function(e){var t=document.createElement("option");t.value=e,t.text=r.getFilterText(e,n),o.appendChild(t)})),o.classList.add("data-selector--active"),o})),l(this,"addPlotly",(function(e){return new c(e,r.chartElement)})),this.chartElement=this.getElement(t)}var t,r,n;return t=e,(r=[{key:"getElement",value:function(e){if(!e)throw new Error("An id to the element or an actual element is required");if("string"==typeof e){var t=document.getElementById(e);if(!t)throw new Error("No element with id ".concat(e));return t}return e}},{key:"getParentElement",value:function(){return this.chartElement.parentElement}},{key:"getFilterText",value:function(e){var t=arguments.length>1&&void 0!==arguments[1]?arguments[1]:{},r=t.labelPrefix,n=t.labelSuffix;return"".concat(r||"").concat(e).concat(n||"")}}])&&u(t.prototype,r),n&&u(t,n),e}(),h=(f=[],s=function(e){var t=function(e){if(!e)throw new Error("An indentifying class for the element is required");return document.querySelectorAll(".".concat(e,":not(.dicharts-handler--active)"))}(e.className);Array.prototype.forEach.call(t,(function(e){e.classList.add("dicharts-handler--active")}))},{addChart:function(e){f.push(e),e.onAdd&&e.onAdd(e),s(e)},getCharts:function(){return f}})}]);
