+++ src/visualisation/library/utils/types.ts
@@ -1,12 +1,14 @@
-interface Loading {
-  show: () => void;
-  hide: () => void;
+import { Layout } from 'plotly.js';
+
+export interface DIChartConfig {
+  className: string;
+  plotly?: DIChartPlotlyConfig;
+  onAdd?: (config: DIChartConfig) => void;
 }
 
-export interface DIChartsInstance {
-  loading: Loading;
-  chartElement: HTMLElement;
-  parentElement: HTMLElement | null;
+export interface DIChartPlotlyConfig {
+  layout: Layout;
+  utils: { [key: string]: any }; // eslint-disable-line @typescript-eslint/no-explicit-any
 }
 
 export interface FilterOptions {
